+++ Sources/PaymentButtons/PaymentButton+ImageAsset.swift
@@ -11,6 +11,9 @@ extension PaymentButton {
     var imageAccessibilityLabel: String {
         // NEXT_MAJOR_VERSION: - To be replaced with translation strings.
 
+        if fundingSource == .venmo {
+            return "Venmo"
+        }
         return fileName.starts(with: "credit") ? "PayPal Credit" : "PayPal"
     }
     
@@ -32,10 +35,17 @@ extension PaymentButton {
 
         case .credit:
             imageAssetString += "credit_"
+
+        case .venmo:
+            imageAssetString += "venmo_"
+
+            if size == .mini {
+                imageAssetString += "monogram_"
+            }
         }
 
         switch color {
-        case .gold, .white, .silver:
+        case .gold, .white, .silver, .venmoBlue:
             imageAssetString += "color"
 
         case .black, .darkBlue: